% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.pmb_analyses}
\alias{is.pmb_analyses}
\title{Is MB Analyses?}
\usage{
is.pmb_analyses(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mb_analyses'
}
