% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.pmb_analysis}
\alias{is.pmb_analysis}
\title{Is MB Analysis?}
\usage{
is.pmb_analysis(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mb_analysis'
}
