% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.pmb_model}
\alias{is.pmb_model}
\title{Is MB Model?}
\usage{
is.pmb_model(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A flag indicating whether the test was positive.
}
\description{
Tests whether x is an object of class 'mb_model'
}
