% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{new_pmb_model}
\alias{new_pmb_model}
\title{Low-level PMB model constructor}
\usage{
new_pmb_model(model, code)
}
\arguments{
\item{model}{An object of class \code{"mb_model"}.}

\item{code}{An object of class \code{"pmb_code"}.}
}
\description{
Low-level PMB model constructor
}
