% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmb-code.R
\name{pmb_code}
\alias{pmb_code}
\title{PMB Code}
\usage{
pmb_code(...)
}
\arguments{
\item{...}{Named strings, unevaluated expressions, or \code{\link[=mb_code]{mb_code()}} objects,
or a single unnamed string, unevaluated expressions, or \code{"pmb_code"} object.
Expressions of the form \code{c(...)} and \code{list(...)} with named dots
are also recognized, for use in \code{\link[=pmb_multi_model]{pmb_multi_model()}}.}
}
\value{
An object inheriting from class \code{"pmb_code"}: a named list of
\code{"mb_code"} objects or quoted expressions.
}
\description{
Captures the model description targeting one or several engine, or written
in the pmbr language that is modeled closely after the NIMBLE language.
}
