% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pmb_model}
\alias{pmb_model}
\title{PMB Model}
\usage{
pmb_model(code, ...)
}
\arguments{
\item{code}{Passed on to \code{\link[=pmb_code]{pmb_code()}}, using \code{{{ }}}.}

\item{...}{Passed on to \code{\link[=model]{model()}}.}
}
\value{
An object inheriting from class \code{"pmb_model"}.
}
\description{
Creates object inherting from class \code{"pmb_model"}.
}
