% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-expr.R
\name{pmb_new_expr}
\alias{pmb_new_expr}
\title{new_expr}
\usage{
pmb_new_expr(object, ...)
}
\arguments{
\item{object}{The object to get for.}

\item{...}{Not used.}
}
\value{
The new_expr
}
\description{
new_expr
}
