% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-expr.R
\name{pmb_set_new_expr}
\alias{pmb_set_new_expr}
\title{Set new_expr}
\usage{
pmb_set_new_expr(object, value)
}
\arguments{
\item{object}{The object to set for.}

\item{value}{The new value of new expr.}
}
\value{
The modified object.
}
\description{
Assignment functions and operators to update \code{new_expr}.
}
