% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmbr-package.R
\docType{package}
\name{pmbr-package}
\alias{pmbr}
\alias{pmbr-package}
\title{pmbr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/pmbr}
  \item \url{https://krlmlr.github.io/pmbr}
  \item Report bugs at \url{https://github.com/krlmlr/pmbr/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
