% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pmb_model}
\alias{pmb_model}
\title{MB Model}
\usage{
pmb_model(
  code,
  ...,
  new_expr = NULL,
  new_expr_vec = getOption("mb.new_expr_vec", FALSE)
)
}
\arguments{
\item{code}{Passed on to \code{\link[=pmb_code]{pmb_code()}}}

\item{...}{These dots are for future extensions and must be empty.}

\item{new_expr}{A string of R code specifying the predictive relationships.}

\item{new_expr_vec}{A flag specifying whether to vectorize the new_expr code.}
}
\value{
An object inheriting from class \code{mb_model}.
}
\description{
Creates object inherting from class mb_model.
}
